--- src/main/java/blackjack/domain/ExceptionHandler.java
@@ -1,6 +1,0 @@
-package blackjack.domain;
-
-@FunctionalInterface
-public interface ExceptionHandler<T> {
-    T execute();
-}
