--- src/main/java/blackjack/domain/Result.java
@@ -1,7 +1,0 @@
-package blackjack.domain;
-
-public enum Result {
-    WIN,
-    LOSE,
-    DRAW
-}
