--- src/main/java/blackjack/domain/card/CardShape.java
@@ -1,8 +1,0 @@
-package blackjack.domain.card;
-
-public enum CardShape {
-    HEART,
-    DIAMOND,
-    SPADE,
-    CLOVER
-}
