+++ src/main/java/domain/Result.java
@@ -1,0 +1,7 @@
+package domain;
+
+public enum Result {
+    WIN,
+    LOSE,
+    DRAW
+}
