+++ src/main/java/domain/card/CardShape.java
@@ -1,0 +1,8 @@
+package domain.card;
+
+public enum CardShape {
+    HEART,
+    DIAMOND,
+    SPADE,
+    CLOVER
+}
