+++ src/main/java/domain/user/Player.java
@@ -1,0 +1,14 @@
+package domain.user;
+
+import domain.card.Card;
+import domain.GamePoint;
+
+public interface Player {
+
+    void draw(Card card);
+
+    boolean canReceive();
+
+    GamePoint getGamePoint();
+
+}
