+++ src/main/java/template/ExceptionHandler.java
@@ -1,0 +1,6 @@
+package template;
+
+@FunctionalInterface
+public interface ExceptionHandler<T> {
+    T execute();
+}
