+++ src/main/java/template/Repeater.java
@@ -1,0 +1,16 @@
+package template;
+
+public class Repeater {
+
+    public static <T> T repeat(ExceptionHandler<T> exceptionHandler) {
+        T result = null;
+        while (result == null) {
+            try {
+                result = exceptionHandler.execute();
+            } catch (Exception exception) {
+                System.out.println(exception.getMessage());
+            }
+        }
+        return result;
+    }
+}
