--- src/test/java/blackjack/domain/BlackJackTest.java
@@ -1,13 +1,0 @@
-package blackjack.domain;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-public class BlackJackTest {
-
-    @Test
-    @DisplayName("")
-    void Test() {
-
-    }
-}
