+++ src/test/java/domain/BlackJackTest.java
@@ -1,0 +1,13 @@
+package domain;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+public class BlackJackTest {
+
+    @Test
+    @DisplayName("")
+    void Test() {
+
+    }
+}
